refactor: Season 패키지 변경 및 Game 공통 요소 추가
