+++ soccer-domain/src/main/kotlin/com/example/soccerdomain/game/Game.kt
@@ -1,5 +1,6 @@
 package com.example.soccerdomain.game
 
+import com.example.soccerdomain.team.domain.Team
 import jakarta.persistence.*
 
 
@@ -7,6 +8,13 @@ import jakarta.persistence.*
 @Inheritance(strategy = InheritanceType.SINGLE_TABLE)
 @DiscriminatorColumn
 open class Game(
+
+    @Column(name = "status")
+    @Enumerated(EnumType.STRING)
+    val status: GameStatus = GameStatus.BEFORE,
+
+    @Transient
+    val teams: MutableList<Team> = mutableListOf(),
 
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "game_id")
