+++ soccer-domain/src/main/kotlin/com/example/soccerdomain/game/GameStatus.kt
@@ -1,6 +1,6 @@
 package com.example.soccerdomain.game
 
-enum class LeagueGameStatus {
+enum class GameStatus {
 
     BEFORE,
     ING,
